ref: integration test restructure
