+++ src/test/java/com/_up/megastore/integrations/state/OrderStateTest.java
@@ -1,4 +1,4 @@
-package com._up.megastore.integrations.implementations;
+package com._up.megastore.integrations.state;
 
 import com._up.megastore.controllers.requests.CancelOrderRequest;
 import com._up.megastore.controllers.requests.CreateOrderRequest;
@@ -18,45 +18,12 @@ import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
-import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-class OrderControllerTest extends BaseIntegrationTest {
+public class OrderStateTest extends BaseIntegrationTest {
 
     @MockBean
     private EmailService emailService;
-
-    @Test
-    @Sql("/scripts/orders/save_order.sql")
-    void saveOrder() throws Exception {
-        CreateOrderRequest request = new CreateOrderRequest(
-                UUID.fromString("58fae25b-ea38-4e7b-ab2d-9f555a67836b"), List.of(
-                new OrderDetailRequest(
-                        UUID.fromString("183f205a-3430-4e11-8bca-57672a0ce3ff"),
-                        5
-                ),
-                new OrderDetailRequest(
-                        UUID.fromString("22ede130-726f-49ac-9564-d783fc22a6fa"),
-                        3
-                )
-        )
-        );
-
-        String response = mockMvc.perform(
-                        post("/api/v1/orders")
-                                .contentType(MediaType.APPLICATION_JSON)
-                                .content(toJson(request))
-                ).andDo(print())
-                .andExpect(status().isCreated())
-                .andExpect(jsonPath("$.total").value(80000.0))
-                .andReturn()
-                .getResponse()
-                .getContentAsString();
-
-        assertContains(response, "clientName", "Client Name");
-        assertContains(response, "state", "IN_PROGRESS");
-    }
 
     @Test
     @Sql("/scripts/orders/save_order.sql")
@@ -211,5 +178,4 @@ class OrderControllerTest extends BaseIntegrationTest {
 
         assertContains(response, "message", State.CANCELLED.exceptionMessage);
     }
-
 }
