+++ src/test/resources/scripts/categories/save_and_update_categories.sql
